- Use %zu when formatting size_t. Remove "static_cast <unsigned long> ()" calls.
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@13101 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/src/find_replace.cpp
@@ -1547,7 +1547,7 @@ int FindReplace::FindInFiles(cbFindReplaceData* data)
     }
     else
     {
-        const wxString msg = wxString::Format(_("\"%s\" not found in %lu files"), data->findText, static_cast<unsigned long>(filesList.GetCount()));
+        const wxString msg = wxString::Format(_("\"%s\" not found in %zu files"), data->findText, filesList.GetCount());
         LogSearch(_T(""), -1, msg );
         m_pSearchLog->FocusEntry(oldcount);
